requirements.txt
fastapi
uvicorn
httpx
python-dotenv
pydantic
boto3
botocore
requests
python-jose
python-multipart
